Removes user prompts and adds arbitrary injection
This version doesn't ask the user to say anything to work. In addition, it has an experimental change where the function.txt file lacking the virtual address to the original function that was called simply causes it to not perform that call. The call is replaced with NOPs, such that the code returns in place. This potentially enables you to inject into any location, execute code, then return to normal code without having to overwrite an existing call. You do have to have 5 bytes to safely kill, but if you do, it should work.